feat(test): adjusting CDP electron runner
