+++ tests/playwright/resources/kube-config
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: Config
+clusters: []
+users: []
+contexts: []
+preferences: {}
+current-context: ''
